requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
bidict==0.23.1
channels==4.1.0
channels-redis==4.2.0
click==8.1.7
Django==5.1.1
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
h11==0.14.0
idna==3.10
msgpack==1.1.0
multidict==6.1.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.0.0
python-engineio==4.9.1
python-socketio==5.11.4
redis==5.0.8
simple-websocket==1.0.0
sqlparse==0.5.1
uvicorn==0.30.6
wsproto==1.2.0
yarl==1.11.1
